Use a distinct name for a field which will error out, not just `error`.
This has confused me like 40 times while looking at this file which has
`trace.root.error` properties which are not the field itself, they are an
actual property on the `Trace.INode` which is called `error`.

This just changes the name of the field, which is actually "error", to be
something that's more unique.